Add not found error

+++ todo-backend/app/Http/Controllers/TodoController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Helpers\ResponseStatus;
 use App\Models\Todo;
 use Illuminate\Http\Request;
 
@@ -12,8 +13,10 @@ class TodoController extends Controller
         return format_json_response(['items' => Todo::all()]);
     }
  
-    public function show(Todo $todo)
+    public function show($id)
     {
+        $todo = Todo::find($id);
+        if (is_null($todo)) return format_json_response((object)[], ['Cannot find todo'], ResponseStatus::NotFound);
         return format_json_response($todo);
     }
 
@@ -24,15 +27,19 @@ class TodoController extends Controller
         return format_json_response($todo, ['Todo added']);
     }
 
-    public function update(Request $request, Todo $todo)
+    public function update(Request $request, $id)
     {
+        $todo = Todo::find($id);
+        if (is_null($todo)) return format_json_response((object)[], ['Cannot find todo'], ResponseStatus::NotFound);
         $todo->update($request->all());
 
         return format_json_response($todo, ['Todo updated']);
     }
 
-    public function delete(Request $request, Todo $todo)
+    public function delete(Request $request, $id)
     {
+        $todo = Todo::find($id);
+        if (is_null($todo)) return format_json_response((object)[], ['Cannot find todo'], ResponseStatus::NotFound);
         $todo->delete();
 
         return format_json_response($todo, ['Todo deleted']);
